flags: Compute OF/sub based on comparison
Previously, the overflow flag for subtraction was computing using
bitwise arithmetic, causing problems for the instruction combiner to
properly fold conditional instructions with the correct predicate.

By computing the overflow flag using a signed less-than comparison, we
not only save an LLVM-IR instruction but also allow for simple folding
of signed comparison conditions.