lifter: Support instruction overriding
Allow to set a custom LLVM function as handler for an instruction type,
overriding the default lifter. The LLVM function takes the CPU state as
single argument. The decoded instruction is not forwarded, because this
would require to store that data in the generated code.

Further optimizations can be performed using passes after lifting, e.g.
inlining and constant propagation.

This allows tools to override the default lifting of specific
instructions, e.g. the SYSCALL or INT3 instructions, for special
handling.