callconv/hhvm: No longer use r14
While increasing the number of registers that are forcibly stored in
memory is favorable, applications have only very few spare registers to
use. This causes even higher overhead, because the register must be
transferred to/from memory all the time, whereas otherwise the value is
only loaded when it is actually required.

No longer storing the guest register R12 in host R14 makes one register
entirely available for other use.