general: Support lifting call/ret to call/ret
This adds two new configuration options: a function to use for tail
continuations and a function to use for function calls.

When a tail continuation function is specified, the exit block does not
return but performs a `musttail` call to that function. Tail
continuation is only performed once in the exit block.

When a call function is given, a CALL instruction will generate a call
to that function (setting RIP accordingly). The CFG decoder will assume
that the function is likely to return and continue decoding after a CALL
instruction. However, it will verify that the RIP matches and otherwise
return/tail continue. When CALL is lifted to an LLVM-IR call, a RET will
be forcefully lifted as LLVM-IR return, regardless of whether tail
continuation is enabled. This may lead to multiple returns in addition
to the single exit block.