lifter: Change ZF computation for SUB
Originally, the zero flag was set if the result of the subtraction (for
SUB/CMP instructions) was zero. This led to an optimization barrier for
comparisons (and useless subtractions in the compiler output); for which
reason it was changed to be set when the two operands are equal.

While this works well for CMP, it causes a substraction _and_ a
comparison to be emitted for SUB instructions. Therefore, we revert to
the old behavior for SUB.