a64: Add TypeOf, Addr and implement LDR, STR
Atomic ordering is still missing, but the bulk of LDRs and STRs are
implemented and tested.

To prioritise the right instructions, I took a Go 1.14 binary of 17
megabytes as the basis for a statistical breakdown. The top 10
most common instructions were:

	Instr.   Frequency (%)
	LDR         23.8
	STR         21.1
	ADD_IMM      7.6
	BL           6.2
	ADRP         4.6
	B.cond       4.1
	B            4.0
	CMP_SHIFTED  3.0
	MOV_REG      2.9
	MOV_IMM      2.0

This means we can already decode 52.5% of this large binary (LDR,
STR, ADD_IMM).