a64: Add LD2/ST2, Load/StoreMulti, common shufflemasks
There are many overlaps between the various LDx/STx sizes, so add
the LoadMulti, StoreMulti, SIMDLoadStoreAddr helpers. This was also
a good opportunity to extract some common shufflemasks into even(),
odd(), zipmask_lower(), zipmask_higher().

The code generated for LD2 is equivalent to the LLVM test case for
lowering to LD2, which is lovely. Onward to LD3/ST3, LD4/ST4, which
require a few more shuffles!