a64: Add multiple-structures LD3/ST3, LD4/ST4
For the loads, the LLVM-IR lowers to the equivalent load, according
to my Godbolt test. Sadly, the same does not hold for the stores.
I believe this is as good as it gets, as LLVM's own code generator
does not lower to STx unless given the appropriate intrinsics
(e.g., llvm.aarch64.neon.st4.v4f32.p0f32).