x86-64: Compute parity lazily
The parity flag is rarely used, but LLVM's lowering makes the
computation fairly expensive. Delay the computation of the parity value
to its actual use by not storing the parity value, but the byte of which
the parity value is computed. This has the minor inconvinience that
value 0 implies PF=1, which slighly complicates the test driver.